backend/requirements.txt
fastapi[all]>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
python-multipart>=0.0.9
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-fuzzy>=0.4.2
loguru>=0.7.0
paho-mqtt>=1.6.0
pandas>=2.0.0
scikit-image>=0.22.0
requests>=2.32.0
scikit-learn>=1.7.0
